refactor: split button types into separate files
- Created src/buttonTypes/ directory with individual files for each button type
- Added registry system to handle circular dependencies
- Updated imports to use new modular structure
- Added setup system for registering button handlers
- Fixed TypeScript configuration to support esModuleInterop
- All button types now have their own dedicated files:
  - calculate.ts
  - chain.ts
  - command.ts
  - copy.ts
  - link.ts
  - remove.ts
  - replace.ts
  - swap.ts
  - template.ts
  - templater.ts
  - text.ts
- Maintained backward compatibility through index.ts exports
