+++ src/buttonTypes/link.ts
@@ -1,0 +1,6 @@
+import { Arguments } from "../types";
+
+export const link = ({ action }: Arguments): void => {
+  const link = action.trim();
+  window.open(link);
+};
